Added initial web idl parser code.
After trying to implement the DOM by hand
I realized that most of the C++ DOM needs to be scrapped
in order to better follow the standard and automate bindings
rather than making a mess of unmanageable code,
so here I am, implementing the webidl parser on Bison and Flex
because Chromium's is Python and not something I could just
copy paste or somehow depend on by the usual means.

PS: This compiles but has not been tested as should be obvious
by the empty main function in the webidl executable.